# Conflicts:
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/AccountController.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/EditAccountController.java
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/utility/DateUtility.java
#	portfolio/src/main/resources/templates/editAccount.html

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/SprintLifetimeController.java
@@ -1,8 +1,13 @@
 package nz.ac.canterbury.seng302.portfolio.controller;
 
 import nz.ac.canterbury.seng302.portfolio.model.Sprint;
+import nz.ac.canterbury.seng302.portfolio.service.ElementService;
+import nz.ac.canterbury.seng302.portfolio.service.PermissionService;
 import nz.ac.canterbury.seng302.portfolio.service.SprintService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,6 +22,15 @@ public class SprintLifetimeController {
 
     @Autowired
     private SprintService sprintService;
+
+    @Autowired
+    private PermissionService permissionService;
+
+    @Autowired
+    private UserAccountClientService userAccountClientService;
+
+    @Autowired
+    private ElementService elementService;
 
     /**
      * Add a given number of days and/or weeks to a date.util using Calendars.
@@ -40,11 +54,18 @@ public class SprintLifetimeController {
      */
     @PostMapping("/add-sprint")
     public String projectSave(
-            @ModelAttribute("sprint") Sprint sprint
+            @ModelAttribute("sprint") Sprint sprint,
+            @AuthenticationPrincipal AuthState principal,
+            Model model
     ) {
-        sprint.setStartDateString(sprint.getStartDateString());
-        sprint.setEndDateString(sprint.getEndDateString());
-        sprintService.addSprint(sprint);
+        Integer userID = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, userID);
+        if (permissionService.isValidToModifyProjectPage(userID)) {
+            sprint.setStartDateString(sprint.getStartDateString());
+            sprint.setEndDateString(sprint.getEndDateString());
+            sprintService.addSprint(sprint);
+        }
+
         return "redirect:/details";
     }
 
@@ -54,10 +75,14 @@ public class SprintLifetimeController {
      * @throws Exception If deleting sprint does not work
      */
     @GetMapping("/delete-sprint/{id}")
-    public String sprintRemove(@PathVariable("id") Integer id, Model model) throws Exception {
-
-        sprintService.removeSprint(id);
-
+    public String sprintRemove(@PathVariable("id") Integer id,
+                               @AuthenticationPrincipal AuthState principal,
+                               Model model) throws Exception {
+        Integer userID = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, userID);
+        if (permissionService.isValidToModifyProjectPage(userID)) {
+            sprintService.removeSprint(id);
+        }
         /* Return the name of the Thymeleaf template */
         return "redirect:/details";
     }
